+++ src/lib/prisma.ts
@@ -1,12 +1,9 @@
 import { PrismaClient } from '@prisma/client'
 import { PrismaLibSql } from '@prisma/adapter-libsql'
-import { createClient } from '@libsql/client'
 
-const libsql = createClient({
+const adapter = new PrismaLibSql({
     url: 'file:./dev.db',
 })
-
-const adapter = new PrismaLibSql(libsql)
 
 const globalForPrisma = globalThis as unknown as { prisma: PrismaClient }